test(server): finish issue 197 and 198 test splits
