+++ nodelite-server/src/audit/mod.rs
@@ -1,8 +1,14 @@
 //! Security audit log module wiring and re-exports.
 
+#[cfg(test)]
+mod disabled_tests;
 mod log;
 mod query;
 mod storage;
+#[cfg(test)]
+mod storage_tests;
+#[cfg(test)]
+mod support;
 #[cfg(test)]
 mod tests;
 mod types;
