+++ nodelite-server/src/audit/tests.rs
@@ -1,34 +1,9 @@
-use std::path::PathBuf;
-use std::time::{SystemTime, UNIX_EPOCH};
-
 use chrono::{Duration as ChronoDuration, Utc};
 use serde_json::json;
 use tokio::runtime::Runtime;
 
-use super::{AuditEventType, AuditLog, AuditLogError, AuditQuery, NewAuditEvent};
-
-#[cfg(unix)]
-use std::os::unix::fs::PermissionsExt;
-
-fn unique_temp_dir(prefix: &str) -> PathBuf {
-    let unique = SystemTime::now()
-        .duration_since(UNIX_EPOCH)
-        .expect("clock should be monotonic enough")
-        .as_nanos();
-    std::env::temp_dir().join(format!("{prefix}-{unique}"))
-}
-
-fn sample_config(db_path: PathBuf) -> nodelite_proto::AuditConfig {
-    nodelite_proto::AuditConfig {
-        enabled: true,
-        db_path,
-        retention_days: 90,
-        log_successful_auth: true,
-        log_failed_auth: true,
-        log_token_events: true,
-        log_rate_limit: true,
-    }
-}
+use super::{AuditEventType, AuditLog, AuditQuery, NewAuditEvent};
+use crate::audit::support::{sample_config, unique_temp_dir};
 
 #[test]
 fn audit_log_round_trips_and_filters_events() {
@@ -253,75 +228,5 @@ fn audit_log_drains_burst_writes_through_writer_task() {
 
         let _ = std::fs::remove_file(&db_path);
         let _ = std::fs::remove_dir_all(&temp_dir);
-    });
-}
-
-#[test]
-#[cfg(unix)]
-fn audit_database_artifacts_are_mode_600() {
-    let runtime = Runtime::new().expect("runtime should build");
-    runtime.block_on(async {
-        let temp_dir = unique_temp_dir("nodelite-audit-mode");
-        std::fs::create_dir_all(&temp_dir).expect("temp dir should exist");
-        let db_path = temp_dir.join("audit.sqlite3");
-        let audit = AuditLog::new(sample_config(db_path.clone()), 5);
-        audit.initialize().await.expect("audit should initialize");
-        audit
-            .record(NewAuditEvent::now(
-                AuditEventType::NodeConnected,
-                "198.51.100.20",
-                true,
-            ))
-            .await
-            .expect("audit event should persist");
-        audit.shutdown().await;
-
-        let data_dir_mode = std::fs::metadata(&temp_dir)
-            .expect("temp dir metadata")
-            .permissions()
-            .mode()
-            & 0o777;
-        assert_eq!(data_dir_mode, 0o700);
-
-        let db_mode = std::fs::metadata(&db_path)
-            .expect("db metadata")
-            .permissions()
-            .mode()
-            & 0o777;
-        assert_eq!(db_mode, 0o600);
-
-        let _ = std::fs::remove_file(&db_path);
-        let _ = std::fs::remove_dir_all(&temp_dir);
-    });
-}
-
-#[test]
-fn disabled_audit_log_rejects_queries_but_ignores_records() {
-    let runtime = Runtime::new().expect("runtime should build");
-    runtime.block_on(async {
-        let mut config = sample_config(PathBuf::from("/tmp/disabled-audit.sqlite3"));
-        config.enabled = false;
-        let audit = AuditLog::new(config, 5);
-
-        audit
-            .record(NewAuditEvent::now(
-                AuditEventType::LoginFailure,
-                "127.0.0.1",
-                false,
-            ))
-            .await
-            .expect("disabled audit log should no-op on record");
-
-        let error = audit
-            .query(AuditQuery {
-                start: None,
-                end: None,
-                event_type: None,
-                success: None,
-                limit: 10,
-            })
-            .await
-            .expect_err("disabled audit log should reject queries");
-        assert!(matches!(error, AuditLogError::Disabled));
     });
 }
