+++ nodelite-server/src/tests/protected_headers_tests.rs
@@ -12,8 +12,10 @@ use axum::routing::get;
 use tokio::runtime::Runtime;
 use tower::util::ServiceExt;
 
+use super::support::{
+    assert_security_headers, json_request, json_write_routes, small_json_write_requests,
+};
 use super::{AppState, set_protected_response_headers};
-use super::support::{assert_security_headers, json_request, json_write_routes, small_json_write_requests};
 use crate::handlers::{index, logout_and_reauth, require_readonly_auth, verify_2fa_page};
 use crate::test_support::{test_server_config, test_ws_config};
